requirements.txt
Flask==3.0.0
gunicorn
mysql-connector-python==8.1.0
python-dotenv==1.0.0